lsf-scraper/requirements.txt
aiohttp==2.3.6
astroid==1.5.3
async-timeout==2.0.0
asyncio==3.4.3
beautifulsoup4==4.6.0
bs4==0.0.1
certifi==2017.11.5
chardet==3.0.4
click==6.7
Flask==0.12.2
gunicorn==19.7.1
icalendar==4.0.0
idna==2.6
isort==4.2.15
itsdangerous==0.24
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.0
mccabe==0.6.1
multidict==3.3.2
pylint==1.7.4
python-dateutil==2.6.1
pytz==2017.3
requests==2.18.4
six==1.11.0
urllib3==1.22
Werkzeug==0.12.2
wrapt==1.10.11
yarl==0.15.0
